# openapi.yaml
openapi: 3.0.3
info:
  title: Brain Agriculture API
  version: 1.0.0
  description: >
    Cadastro de produtores rurais, fazendas, plantios por safra e dashboard JSON.
    Sem UI Swagger. Base `/api/v1`; health fora do prefixo.
servers:
  - url: http://localhost:3000
    description: Local
tags:
  - name: Health
  - name: Produtores
  - name: Fazendas
  - name: Plantios
  - name: Dashboard
paths:
  /health:
    get:
      tags: [Health]
      summary: Liveness
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
  /api/v1/producers:
    get:
      tags: [Produtores]
      summary: Lista paginada (sem fazendas)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Lista
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProducerList'
    post:
      tags: [Produtores]
      summary: Cria produtor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProducer'
            example:
              name: João da Silva
              document: '529.982.247-25'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producer'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/producers/{id}:
    get:
      tags: [Produtores]
      summary: Detalhe com fazendas (sem plantios)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Produtor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProducerWithFarms'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Produtores]
      summary: Atualiza nome e/ou documento
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProducer'
      responses:
        '200':
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producer'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Produtores]
      summary: Remove produtor sem fazendas
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Removido
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /api/v1/producers/{id}/farms:
    post:
      tags: [Fazendas]
      summary: Cria fazenda do produtor
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFarm'
            example:
              name: Fazenda Boa Vista
              city: Ribeirão Preto
              state: SP
              totalAreaHa: 100
              arableAreaHa: 60
              vegetationAreaHa: 30
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/farms/{id}:
    get:
      tags: [Fazendas]
      summary: Detalhe com plantios
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Fazenda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmDetail'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Fazendas]
      summary: Atualiza fazenda (recalcula áreas no estado final)
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFarm'
      responses:
        '200':
          description: Atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Fazendas]
      summary: Remove fazenda e plantios (cascade)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Removida
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/farms/{id}/plantings:
    post:
      tags: [Plantios]
      summary: Cria plantio (get-or-create safra e cultura)
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanting'
            example:
              harvestName: Safra 2021
              cropName: Soja
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Planting'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /api/v1/plantings/{id}:
    delete:
      tags: [Plantios]
      summary: Remove só o plantio
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Removido
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/dashboard:
    get:
      tags: [Dashboard]
      summary: Agregados para pizza (JSON, sem gráfico)
      responses:
        '200':
          description: Snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: RFC 7807
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, requestId]
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        requestId: { type: string }
    CreateProducer:
      type: object
      required: [name, document]
      properties:
        name: { type: string, minLength: 2, maxLength: 255 }
        document: { type: string, description: 'CPF ou CNPJ, com ou sem máscara' }
    UpdateProducer:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 255 }
        document: { type: string }
    Producer:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        document: { type: string, description: 'Somente dígitos' }
        documentType: { type: string, enum: [CPF, CNPJ] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FarmSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        city: { type: string }
        state: { type: string, example: SP }
        totalAreaHa: { type: number }
        arableAreaHa: { type: number }
        vegetationAreaHa: { type: number }
    ProducerWithFarms:
      allOf:
        - $ref: '#/components/schemas/Producer'
        - type: object
          properties:
            farms:
              type: array
              items:
                $ref: '#/components/schemas/FarmSummary'
    ProducerList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Producer'
        meta:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
    CreateFarm:
      type: object
      required: [name, city, state, totalAreaHa, arableAreaHa, vegetationAreaHa]
      properties:
        name: { type: string }
        city: { type: string }
        state: { type: string, description: 'UF IBGE, 27 valores inclusive DF' }
        totalAreaHa: { type: number }
        arableAreaHa: { type: number }
        vegetationAreaHa: { type: number }
    UpdateFarm:
      type: object
      properties:
        name: { type: string }
        city: { type: string }
        state: { type: string }
        totalAreaHa: { type: number }
        arableAreaHa: { type: number }
        vegetationAreaHa: { type: number }
    Farm:
      allOf:
        - $ref: '#/components/schemas/FarmSummary'
        - type: object
          properties:
            producerId: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    FarmPlanting:
      type: object
      properties:
        id: { type: string, format: uuid }
        harvestName: { type: string }
        cropName: { type: string }
    FarmDetail:
      allOf:
        - $ref: '#/components/schemas/Farm'
        - type: object
          properties:
            plantings:
              type: array
              items:
                $ref: '#/components/schemas/FarmPlanting'
    CreatePlanting:
      type: object
      required: [harvestName, cropName]
      properties:
        harvestName: { type: string }
        cropName: { type: string }
    Planting:
      type: object
      properties:
        id: { type: string, format: uuid }
        farmId: { type: string, format: uuid }
        harvestName: { type: string }
        cropName: { type: string }
        createdAt: { type: string, format: date-time }
    Dashboard:
      type: object
      properties:
        totalFarms: { type: integer }
        totalHectares: { type: number }
        farmsByState:
          type: array
          items:
            type: object
            properties:
              state: { type: string }
              count: { type: integer }
        cropsPlanted:
          type: array
          items:
            type: object
            properties:
              crop: { type: string }
              count: { type: integer }
        landUse:
          type: object
          properties:
            arableHectares: { type: number }
            vegetationHectares: { type: number }